Post article without image too

+++ src/controllers/article.ts
@@ -6,13 +6,18 @@ import s3 from '../storage/aws-s3'
 import Article from '../models/article'
 
 export const postArticle: RequestHandler = async (req: Request, res: Response) => {
-  const imgFile = <File> req.file
   const { articleTitle, articleBody } = req.body
-  const imageName = imgFile.originalname
 
+  let imgFile: File;
+  let imageName = undefined;
+  let s3Res;
   try {
-    const s3Res = await s3.saveFile(imgFile)
-    fs.unlinkSync(imgFile.path)
+    if (req.file) {
+      imgFile = <File> req.file
+      imageName = imgFile.originalname
+      s3Res = await s3.saveFile(imgFile)
+      fs.unlinkSync(imgFile.path)
+    }
 
     const article = new Article({
       title: articleTitle,
@@ -25,7 +30,7 @@ export const postArticle: RequestHandler = async (req: Request, res: Response) =
       code: 'POSTED',
       msg: {
         db: `Document inserted with _id: ${dbRes.insertedId}`,
-        s3: `Image uploaded with ETag: ${s3Res.ETag}`
+        s3: s3Res ? `Image uploaded with ETag: ${s3Res.ETag}` : 'NO_IMAGE'
       }
     })
